updated post controller a user can delete any comment on his post

+++ routers/controller/post.js
@@ -152,24 +152,28 @@ const deletePostComment = (req, res) => {
     .then((postResult) => {
       if (postResult.userId == req.addedToken.id) {
         //يقارن اليوزر ايدي عند البوست باليوزر ايدي بالبايلود
-        console.log(postResult.userId);
-        console.log(req.addedToken.id);
-        console.log("first If");
+        // console.log(postResult.userId);
+        // console.log(req.addedToken.id);
+        // console.log("first If");
         commentModel.findById({ _id: commentId }).then((comResult) => {
-          console.log(postResult._id);
-          console.log(comResult.postId);
-          console.log("second If");
-          if (postResult._id.toString() == comResult.postId.toString()) {
-            //يقارن البوست ايدي بالبوست ايدي بالكومنت
-            console.log("you'r in!!");
-            commentModel
-              .findByIdAndDelete({ _id: commentId })
-              .then((result) => {
-                console.log("last round");
-                res.status(200).json("comment have been deleted");
-              });
+          if (comResult) {
+            // console.log(postResult._id);
+            // console.log(comResult.postId);
+            // console.log("second If");
+            if (postResult._id.toString() == comResult.postId.toString()) {
+              //يقارن البوست ايدي بالبوست ايدي بالكومنت
+              console.log("you'r in!!");
+              commentModel
+                .findByIdAndDelete({ _id: commentId })
+                .then((result) => {
+                  console.log("last round");
+                  res.status(200).json("comment have been deleted");
+                });
+            } else {
+              res.status(400).send("you are not allowed to delete it :/");
+            }
           } else {
-            res.status(400).send("you are not allowed to delete it :/");
+            res.status(400).send("comment not found");
           }
         });
       } else {
